test(README): add test for usage code block
Change the type of the README usage code block from INI to TOML, so that
it gets picked up by the version-sync test.
Add `version-sync` library as a dev dependency.
Add a version-sync integration test for the README usage code block to
ensure that the code block never gets out-of-date.